+++ tools/evm-toolkit/src/lib/iscmagic/utils.ts
@@ -1,4 +1,8 @@
+import { Blake2b } from '@iota/crypto.js';
 import { Converter } from '@iota/util.js';
+import { Buffer } from 'buffer';
+
+import { SimpleBufferCursor } from '$lib/simple-buffer-cursor';
 
 export function evmAddressToAgentID(evmStoreAccount: string): Uint8Array {
   // This function constructs an AgentID that is required to be used with contracts
@@ -13,4 +17,19 @@ export function evmAddressToAgentID(evmStoreAccount: string): Uint8Array {
   ]);
 
   return addressBytes;
+}
+
+export function hNameFromString(name): number {
+  const ScHNameLength = 4;
+  const stringBytes = Converter.utf8ToBytes(name);
+  const hash = Blake2b.sum256(stringBytes);
+
+  for (let i = 0; i < hash.length; i += ScHNameLength) {
+    const slice = hash.slice(i, i + ScHNameLength);
+    const cursor = new SimpleBufferCursor(Buffer.from(slice));
+
+    return cursor.readUInt32LE();
+  }
+
+  return 0;
 }
