+++ tools/evm-toolkit/src/lib/nft/constants.ts
@@ -1,0 +1,3 @@
+import type { INFTMetaDataCacheMap } from './interfaces';
+
+export const NFTMetadataCache: INFTMetaDataCacheMap = {};
